pyproject.toml
[tool.poetry]
name = "paysage-classification"
version = "0.2.3"
description = "Landscape image classification using EfficientNetV2"
authors = ["zhabiki team"]
readme = "README.MD"

[tool.poetry.dependencies]
bottleneck = "^1.4.0"
dvc = "^3.66.1"
fsspec = "^2025.10.0"
hydra-core = "^1.3.2"
jinja2 = "^3.1.0"
joblib = "^1.5.0"
matplotlib = "^3.10.0"
mlflow = "^3.8.1"
networkx = "^3.5.0"
numpy = "^2.0.0"
omegaconf = "^2.3.0"
pandas = "^2.3.0"
pillow = "^11.0.0"
python = "^3.11"
python-slugify = "^8.0.4"
pyyaml = "^6.0.0"
requests = "^2.32.0"
scikit-learn = "^1.7.0"
scipy = "^1.16.0"
tensorboard = "^2.20.0"
threadpoolctl = "^3.6.0"
tk = "^0.1.0"
torch = "^2.5.0"
torchaudio = "^2.5.0"
torchvision = "^0.20.0"
tqdm = "^4.67.0"

[tool.poetry.group.dev.dependencies]
black = "^24.0.0"
mypy = "^1.10.0"
pre-commit = ">=3.7.1"
pytest = "^8.3.0"
# Конфиг для flake8 -- в setup.cfg

[tool.black]
line-length = 88
target-version = ["py311"]
skip-string-normalization = true # А мне 'одинарные кавычки' дороже автоформатирования!

[tool.isort]
profile = "black"
line_length = 88
skip_gitignore = true

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = false
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
